+++ src/style/load_sprite.ts
@@ -55,9 +55,9 @@ async function doOnceCompleted(
         const json = (await jsonsMap[spriteName]).data;
 
         for (const id in json) {
-            const {width, height, x, y, sdf, pixelRatio, stretchX, stretchY, content} = json[id];
+            const {width, height, x, y, sdf, pixelRatio, stretchX, stretchY, content, textFitWidth, textFitHeight} = json[id];
             const spriteData = {width, height, x, y, context};
-            result[spriteName][id] = {data: null, pixelRatio, sdf, stretchX, stretchY, content, spriteData};
+            result[spriteName][id] = {data: null, pixelRatio, sdf, stretchX, stretchY, content, textFitWidth, textFitHeight, spriteData};
         }
     }